minor modification

+++ src/iostar_json/show_configuration.py
@@ -10,14 +10,14 @@ class ShowConfiguration:
     nb_y: int  # Number of families on the y-axis during the takeoff
     nb_drone_per_family: int  # Number of drones in each families
     step: int  # Distance separating the families during the takeoff in centimeter
-    angle_takeoff: int  # Angle of the takeoff grid
-    duration: int  # Duration of the show in millisecond
+    angle_takeoff: int  # Angle of the takeoff grid in degree
+    duration: float  # Duration of the show in second
     hull: List[
-        Tuple[int, int]
-    ]  # List of the relative coordinate (XY in NED and centimeter) symbolysing a convex hull of a show
+        Tuple[float, float]
+    ]  # List of the relative coordinate (XY in ENU and meter) symbolysing a convex hull of a show
     altitude_range: Tuple[
-        int, int
-    ]  # Relative coordinate ( z_min and z_max in NED and centimeter) symbolising the range of the z-axis
+        float, float
+    ]  # Relative coordinate ( z_min and z_max in ENU and meter) symbolising the range of the z-axis
 
     @property
     def nb_family(self) -> int:
